Add files via upload
You must have heard a game named Housie, I hated that game, the reason is that a person has to pick a random number from 1 - 99 and if the number is in your ticket, its good for you. But sometimes people cheat in this game by partnering with the person who is picking the numbers. Thats why i made this program, so we could play it fair!!